+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Handle invalid JSON parsing in request bodies
 * Fix the request body that keeps changing preventing editing of it.
 * Set min height on syntax rendering to be 2rem, so that the copy button always fully shows correctly.
+* Fix default display when the value is falsy.
 
 ## 2.1
 * Add `x-locale` vendor extension to specify the locale of the spec.